Split process into files. Remove all ASCII win32 functions and replace them with widechar version
